TLRC - 895 - Implement Error Reporting for Failed Uploads (Stage 2 Validations) - TLRC - 1017 - Dev: Web - Redirecting to UploadUnsuccessful page

+++ src/Sfa.Tl.ResultsAndCertification.Web/ViewModel/Registration/UploadUnsuccessfulViewModel.cs
@@ -3,6 +3,6 @@
     public class UploadUnsuccessfulViewModel
     {
         public string FileType { get; set; }
-        public int FileSize {get;set;}
+        public double FileSize {get;set;}
     }
 }
